[PM-15906] Implement single tap passkey flows
This change adds a biometric prompt to the passkey autofill flow.

The biometric prompt will be shown when creating or authenticating with a passkey if the user has supported device biometrics enabled.

This change also adds an isUserVerified flag to the Fido2 requests to determine if the user has verified their identity using the single tap flow.